pta: stats: fix compilation incompatible pointer warning
To reproduce the issue:
$ make PLATFORM=imx-mx8mmevk CFG_WITH_STATS=y CFG_TA_STATS=y

core/pta/stats.c: In function ‘get_user_ta_stats’:
core/pta/stats.c:169:37: warning: passing argument 2 of ‘tee_ta_instance_stats’ from incompatible pointer type [-Wincompatible-pointer-types]
  169 |                                     &p[0].memref.size);
      |                                     ^~~~~~~~~~~~~~~~~
      |                                     |
      |                                     size_t * {aka long unsigned int *}
In file included from core/include/kernel/pseudo_ta.h:10,
                 from core/pta/stats.c:8:
core/include/kernel/tee_ta_manager.h:171:56: note: expected ‘uint32_t *’ {aka ‘unsigned int *’} but argument is of type ‘size_t *’ {aka ‘long unsigned int *’}
  171 | TEE_Result tee_ta_instance_stats(void *buff, uint32_t *buff_size);
      |                                              ~~~~~~~~~~^~~~~~~~~
core/pta/stats.c: At top level:
cc1: note: unrecognized command-line option ‘-Wno-c2x-extensions’ may have been intended to silence earlier diagnostics

Fixes: 7509620b8b95 ("GP131: Update TEE_Param")
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>

+++ core/kernel/tee_ta_manager.c
@@ -966,7 +966,7 @@ static TEE_Result dump_ta_stats(struct tee_ta_dump_ctx *dump_ctx,
 	return TEE_SUCCESS;
 }
 
-TEE_Result tee_ta_instance_stats(void *buf, uint32_t *buf_size)
+TEE_Result tee_ta_instance_stats(void *buf, size_t *buf_size)
 {
 	TEE_Result res = TEE_SUCCESS;
 	struct tee_ta_dump_stats *dump_stats = NULL;
